pyproject.toml
[build-system]
requires = ["hatchling>=1.28.0"]
build-backend = "hatchling.build"

[project]
name = "ductor-slack"
version = "0.16.0"
description = "Control Claude Code and Codex CLI from Telegram. Live streaming, sessions, cron jobs, webhooks, Docker sandboxing."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "PleasePrompto" }]
keywords = [
    "ductor", "ai", "telegram", "bot", "agent",
    "claude", "codex", "cli", "automation", "streaming",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.12.5",
    "aiogram>=3.24.0,<4.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    "cronsim>=2.7",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "questionary>=2.1.1",
    "filetype>=1.2.0",
    "Pillow>=10.0.0",
    "tzdata>=2024.1",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-aiohttp>=1.1.0",
    "pytest-cov>=7.0.0",
    "time-machine>=3.2.0",
    "pytest-xdist>=3.8.0",
]
lint = [
    "ruff>=0.15.0",
    "mypy>=1.19.1",
]
api = [
    "PyNaCl>=1.6.2",
]
matrix = [
    "matrix-nio>=0.25.0",
]
slack = [
    "slack-bolt>=1.23.0,<2.0.0",
    "slack-sdk>=3.37.0,<4.0.0",
]
dev = ["ductor-slack[test,lint,api,matrix,slack]"]

[project.urls]
Homepage = "https://ductor.dev"
Repository = "https://github.com/PleasePrompto/ductor"
Documentation = "https://ductor.dev"
Issues = "https://github.com/PleasePrompto/ductor/issues"
Changelog = "https://github.com/PleasePrompto/ductor/releases"

[project.scripts]
ductor-slack = "ductor_slack.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["ductor_slack"]

[tool.hatch.build.targets.wheel.force-include]
"config.example.json" = "ductor_slack/_config_example.json"
"Dockerfile.sandbox" = "ductor_slack/_Dockerfile.sandbox"

[tool.hatch.build.targets.sdist]
include = [
    "ductor_slack/",
    "config.example.json",
    "Dockerfile.sandbox",
    "LICENSE",
    "README.md",
    "pyproject.toml",
]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = ["ductor_slack/_home_defaults", ".claude"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "E501",
    "W191",
    "COM812",
    "ISC001",
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107",
    "D203",
    "D213",
    "ANN401",
    "TD002", "TD003",
    "FIX002",
    "ERA001",
    "S101",
    "TRY003",
    "EM101", "EM102",
    "T201",
    "FBT001", "FBT002",
    "S603", "S607",
    "SLF001",
    "TC001", "TC003",
    "PLR2004",
    "PLC0415",
    "BLE001",
    "ARG002",
    "PERF203",
    "RUF012",
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"ductor_slack/_home_defaults/**" = ["ALL"]
"**/tests/*" = [
    "S101", "ANN", "D",
    "S105", "S106", "S107", "S108", "S603",  # test tokens/paths/subprocess are not secrets
    "PT019",                          # @patch injections look like unused fixtures to ruff
    "PLR0913", "PLR0915", "PLR0911",  # test helpers need flexible limits
    "FBT003",                         # boolean positional args in patch/assert calls
]

[tool.ruff.lint.isort]
known-first-party = ["ductor_slack"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
disallow_any_generics = true
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_reexport = true
no_implicit_optional = true
strict_equality = true
extra_checks = true

[[tool.mypy.overrides]]
module = ["aiogram.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["cronsim.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["yaml.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["questionary.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["filetype.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["nacl.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["nio.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["slack_bolt.*", "slack_sdk.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["PIL.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["ductor_slack._home_defaults.*"]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
